+++ stdlib/Random/src/normal.jl
@@ -35,7 +35,7 @@ julia> randn(rng, ComplexF32, (2, 3))
   0.611224+1.56403im   0.355204-0.365563im  0.0905552+1.31012im
 ```
 """
-@inline function randn(rng::AbstractRNG=GLOBAL_RNG)
+@inline function randn(rng::AbstractRNG=get_local_rng())
     @inbounds begin
         r = rand(rng, UInt52())
         rabs = Int64(r>>1) # One bit for the sign
@@ -95,7 +95,7 @@ julia> randexp(rng, 3, 3)
  0.695867  0.693292  0.643644
 ```
 """
-function randexp(rng::AbstractRNG=GLOBAL_RNG)
+function randexp(rng::AbstractRNG=get_local_rng())
     @inbounds begin
         ri = rand(rng, UInt52())
         idx = ri & 0xFF
@@ -165,7 +165,7 @@ for randfun in [:randn, :randexp]
     @eval begin
         # scalars
         $randfun(rng::AbstractRNG, T::BitFloatType) = convert(T, $randfun(rng))
-        $randfun(::Type{T}) where {T} = $randfun(GLOBAL_RNG, T)
+        $randfun(::Type{T}) where {T} = $randfun(get_local_rng(), T)
 
         # filling arrays
         function $randfun!(rng::AbstractRNG, A::AbstractArray{T}) where T
@@ -175,19 +175,19 @@ for randfun in [:randn, :randexp]
             A
         end
 
-        $randfun!(A::AbstractArray) = $randfun!(GLOBAL_RNG, A)
+        $randfun!(A::AbstractArray) = $randfun!(get_local_rng(), A)
 
         # generating arrays
         $randfun(rng::AbstractRNG, ::Type{T}, dims::Dims                     ) where {T} = $randfun!(rng, Array{T}(undef, dims))
         # Note that this method explicitly does not define $randfun(rng, T),
         # in order to prevent an infinite recursion.
         $randfun(rng::AbstractRNG, ::Type{T}, dim1::Integer, dims::Integer...) where {T} = $randfun!(rng, Array{T}(undef, dim1, dims...))
-        $randfun(                  ::Type{T}, dims::Dims                     ) where {T} = $randfun(GLOBAL_RNG, T, dims)
-        $randfun(                  ::Type{T}, dims::Integer...               ) where {T} = $randfun(GLOBAL_RNG, T, dims...)
+        $randfun(                  ::Type{T}, dims::Dims                     ) where {T} = $randfun(get_local_rng(), T, dims)
+        $randfun(                  ::Type{T}, dims::Integer...               ) where {T} = $randfun(get_local_rng(), T, dims...)
         $randfun(rng::AbstractRNG,            dims::Dims                     )           = $randfun(rng, Float64, dims)
         $randfun(rng::AbstractRNG,            dims::Integer...               )           = $randfun(rng, Float64, dims...)
-        $randfun(                             dims::Dims                     )           = $randfun(GLOBAL_RNG, Float64, dims)
-        $randfun(                             dims::Integer...               )           = $randfun(GLOBAL_RNG, Float64, dims...)
+        $randfun(                             dims::Dims                     )           = $randfun(get_local_rng(), Float64, dims)
+        $randfun(                             dims::Integer...               )           = $randfun(get_local_rng(), Float64, dims...)
     end
 end