+++ stdlib/Test/src/Test.jl
@@ -27,7 +27,7 @@ export TestSetException
 import Distributed: myid
 
 using Random
-using Random: AbstractRNG, GLOBAL_RNG
+using Random: AbstractRNG, get_local_rng
 using InteractiveUtils: gen_call_with_extracted_types
 using Core.Compiler: typesubtract
 
@@ -1106,10 +1106,11 @@ function testset_beginend(args, tests, source)
         # we reproduce the logic of guardseed, but this function
         # cannot be used as it changes slightly the semantic of @testset,
         # by wrapping the body in a function
-        oldrng = copy(GLOBAL_RNG)
+        local RNG = get_local_rng()
+        oldrng = copy(RNG)
         try
-            # GLOBAL_RNG is re-seeded with its own seed to ease reproduce a failed test
-            Random.seed!(GLOBAL_RNG.seed)
+            # RNG is re-seeded with its own seed to ease reproduce a failed test
+            Random.seed!(RNG.seed)
             $(esc(tests))
         catch err
             err isa InterruptException && rethrow()
@@ -1117,7 +1118,7 @@ function testset_beginend(args, tests, source)
             # error in this test set
             record(ts, Error(:nontest_error, :(), err, Base.catch_stack(), $(QuoteNode(source))))
         finally
-            copy!(GLOBAL_RNG, oldrng)
+            copy!(RNG, oldrng)
         end
         pop_testset()
         finish(ts)
@@ -1176,7 +1177,7 @@ function testset_forloop(args, testloop, source)
             pop_testset()
             push!(arr, finish(ts))
             # it's 1000 times faster to copy from tmprng rather than calling Random.seed!
-            copy!(GLOBAL_RNG, tmprng)
+            copy!(RNG, tmprng)
 
         end
         ts = $(testsettype)($desc; $options...)
@@ -1195,9 +1196,10 @@ function testset_forloop(args, testloop, source)
         arr = Vector{Any}()
         local first_iteration = true
         local ts
-        local oldrng = copy(GLOBAL_RNG)
-        Random.seed!(GLOBAL_RNG.seed)
-        local tmprng = copy(GLOBAL_RNG)
+        local RNG = get_local_rng()
+        local oldrng = copy(RNG)
+        Random.seed!(RNG.seed)
+        local tmprng = copy(RNG)
         try
             $(Expr(:for, Expr(:block, [esc(v) for v in loopvars]...), blk))
         finally
@@ -1206,7 +1208,7 @@ function testset_forloop(args, testloop, source)
                 pop_testset()
                 push!(arr, finish(ts))
             end
-            copy!(GLOBAL_RNG, oldrng)
+            copy!(RNG, oldrng)
         end
         arr
     end
@@ -1648,7 +1650,7 @@ Base.similar(A::GenericArray, s::Integer...) = GenericArray(similar(A.a, s...))
 
 "`guardseed(f)` runs the function `f()` and then restores the
 state of the global RNG as it was before."
-function guardseed(f::Function, r::AbstractRNG=GLOBAL_RNG)
+function guardseed(f::Function, r::AbstractRNG=get_local_rng())
     old = copy(r)
     try
         f()